### 2024-11-12 #7453 @arfon
While we are waiting, could you please:

- Fix the DOIs (noting that @editorialbot's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references:

- Explain how this software is tested? Being able to test the capabilities of software is important (and required for JOSS as part of the review). Please explain here how a user might do this.
- The paper is also a little long (and appears to be missing a _Statement of need_ section), could you please work on that too. Thank you!
