### 2024-11-20 #7453 @kyleniemeyer
@Habiba-Eldababy before we assign an editor, can you please address the question about how your software is tested? We don't require automated testing / continuous integration, but software does need to have tests set up somehow.
