### 2024-11-20 eeaa194
Update paper.bib
Updated DOIs
