+++ Paper/paper.bib
@@ -3,7 +3,8 @@
   title     = {A course on damage mechanics},
   publisher = {Springer},
   year      = {2012},
-  address   = {Berlin}
+  address   = {Berlin},
+  doi       = {10.1007/978-3-642-18255-6}
 }
 
 @article{pebrel2008,
@@ -14,6 +15,7 @@
   number    = {3},
   pages     = {57},
   year      = {2008}
+  doi       = {10.1615/intjmultcompeng.v6.i3.50}
 }
 
 @article{li1997,
