### 2024-11-20 864d9a4
Update paper.md
Renamed "Summary and Statement of Need" to "Statement of Need"
