+++ Paper/paper.bib
@@ -14,7 +14,7 @@
   volume    = {6},
   number    = {3},
   pages     = {57},
-  year      = {2008}
+  year      = {2008},
   doi       = {10.1615/intjmultcompeng.v6.i3.50}
 }
 
