### 2024-12-20 #7610 @editorialbot
[REVIEW]: Parallel Implementation of Continuum Damage Mechanics Simulations using FEM and MATLAB
**Submitting author:** @Habiba-Eldababy (<a href="http://orcid.org/0009-0005-1659-9411">Habiba Eldababy</a>)
**Repository:** https://github.com/Habiba-Eldababy/Parallel-CDM
**Branch with paper.md** (empty if default branch):
**Version:** MATLAB R2022b
