### 2025-01-31 Habiba-Eldababy/Parallel-CDM#1 @tiburoch
JOSS Review: Code features and comparison with other software
Dear @Habiba-Eldababy ,

Your code is geared towards Continuum Damage Mechanics, yet there is no information in the documentation nor the paper about the specific features of your code.

Specifically, which damage models are implemented (local/non-local, specific formulation)? What is the behavior of the bulk? Can you solve 3D problems? What are the 2D conditions  (plane strain/plane stress)? What type of boundary conditions can you use? Is your code solving static or dynamic problems? All those questions (and maybe more) should be addressed in the documentation and the description of your software.

In addition to specifying the features of your code, you should also compare them to the capabilities of other accessible software. What is the main difference of your code compared to existing FEM libraries that include damage mechanics in parallel? For example, I can think of akantu that includes parallel phase-field models (https://joss.theoj.org/papers/10.21105/joss.05253), oofem (https://oofem.org/doku.php?id=en:oofem) that includes both parallelism and many damage models (I am not sure if they work together though...), FEniCS (https://fenicsproject.org/) etc...

This will help clarify what your software does and the statement of needs of the JOSS paper.
