### 2025-01-31 Habiba-Eldababy/Parallel-CDM#2 @tiburoch
JOSS Review: Code documentation and example
Dear @Habiba-Eldababy ,

I believe there is a lack of documentation for your code, which I think is partly related to #1 . It is not clear what is implemented and how a new user can interact with the code apart from changing the parameters in the input files, which are not documented either. To make your code accessible to new users, you should include documentation regarding:

- The minimum structure of a working example, and a description of the corresponding functions. Right now, I would be unable to write a script by myself. How do I read material parameters? How do I apply boundary conditions? How do I call the solver? Etc... As a suggestion, this might benefit from factorizing the code a bit, as there are many variables exposed on the main file, which makes it difficult to read.

- The mesh. What are the supported formats? Maybe you can also suggest a software that generates such mesh format. What types of elements are supported? Can I use an unstructured mesh? Etc...

- The input parameters. Every parameter should be documented. In the case of purely numerical (non-physical) parameters for the solver, can you specify if/how they might impact the validity of the results?

- The content of the outputs.

Documenting all those points will not only allow users to get started with your code, but will also allow them to potentially modify or expand it to their needs.
