### 2025-01-31 Habiba-Eldababy/Parallel-CDM#3 @tiburoch
JOSS Review: Test
Dear @Habiba-Eldababy ,

You need to implement tests for your software, either by comparing a simulation to an analytical solution (if available) or with a benchmark test.
