### 2025-01-31 #7610 @tiburoch
Dear @diehlpk and @Habiba-Eldababy,

I have thoroughly checked the submission. In the current state, I can not recommend this work for publication. Before I can proceed further with the review, important points need to be addressed:

1 - Clarify the features of the software and the statement of needs https://github.com/Habiba-Eldababy/Parallel-CDM/issues/1

2 - Provide documentation for the code, inputs, outputs, and parameters https://github.com/Habiba-Eldababy/Parallel-CDM/issues/2

3 - Write tests https://github.com/Habiba-Eldababy/Parallel-CDM/issues/3

I have opened additional minor issues on the code repository.

Addressing points 1 and 2 might involve significant work and greatly modify the code, documentation and paper. At this stage, other minor comments are not yet relevant.

Best regards,

Thibault Roch
