### 2025-03-24 #7610 @vijaysm
I second many of the review points that @tiburoch has already raised. The repository is missing several recommended and required practices.

1. community contribution guidelines,
2. documentation related to additional dependencies (parallel computing toolbox) and pointers to tools for generating new meshes for test cases,
3. detailed documentation describing the test suite and explaining functionality using tutorials/examples
4. description of drivers FEM_Main_Script_2D.m and FEM_Main_Script_2D_HPC.m, their usage, and along with the differences between the two versions
5. documentation detailing the control of process vs thread-based parallelized runs and methods to change the number of tasks (pointers to MATLAB documentation may suffice along with instructions to run the driver)

I also have a general comment about the paper.

- In Figure 1., what happens when you update the parameters? There is perhaps an arrow missing to restart the iterations?
- In Figure 2., the solver workflow, I only see the assembly of the stiffness matrix being done in parallel. Does the solver use any parallelization? This is not indicated in the flowchart.
- For performance results in Figures 5 and 6, it may be useful to consider your "medium" mesh as the coarse one, and use another level of refinement for the fine one to showcase better results on the workstation and the cluster.
- Is the parallel efficiency of MATLAB thread-level parallelism limited? The scaling plots plateau immediately beyond 2-3 threads on the workstation and 8 threads on a cluster. While there is a reduction in total runtime compared to serial execution, does the author think that leveraging MATLAB's thread-level parallelism provides a speedup comparable to domain-decomposition-based approaches? You can discuss these implications as part of your conclusions.

Please address these suggestions and let me know when the repository/paper are ready to be reviewed again. @Habiba-Eldababy  @diehlpk
