### 2025-04-08 #7610 @diehlpk
@Habiba-Eldababy how is the progress on the open ticket going?
