### 2025-04-26 #7610 @diehlpk
@Habiba-Eldababy how is the progress on the open ticket going?
