### 2025-05-20 a1c78eb
Create ParallelCDM_Documentation.md
