+++ Documentation/ParallelCDM_Documentation.md
@@ -1,0 +1,121 @@
+## I. Software Features
+This code performs parallel-based simulations of continuum damage mechanics problems. In its current formulation, the following features are available:
+
+- The code models the performance of quasi-brittle materials, and the user can select between two relevant damage models that capture such a behavior: the Mazars [1] and the Geers [2] damage laws. 
+- The code performs quasi-static simulations for 2D problems (plane-strain or plane-stress).
+- The code supports the analysis of both structured and unstructured meshes, with quadrilateral elements that contain 4 Gauss points.  
+- Two numerical solvers are available, namely the Newton-Raphson (NR) [3] and the Unified Arc-Length (UAL) [4] methods. 
+- The code supports displacement-driven analysis (Dirichlet BCs are imposed)
+- The code can be run either in serial or parallel mode. 
+
+## II. Software Prerequisites
+This code requires **MATLAB** and the **Parallel Computing Toolbox** to be installed. We use MATLAB version **R2022b**. While we recommend using GMSH to generate a new mesh, the user can pick any relevant software of their choice. The generated mesh should be stored in a ".txt" file, modified by the user to match the format that is explained below (see the “IV. Model-files structure” section). 
+
+## III. Code structure
+The code is a compilation of several ".m" files. The main script is the `FEM_Main_Script_2D.m`, which is the only code-related file the user needs to open and edit. In this file the user specifies the <ins>working/saving directories</ins> (line 13) and the <ins>model name</ins> (line 16). This file then calls the solver (Newton-Raphson or UAL) and the analysis starts. 
+
+Each file named `func____.m` serves a specific purpose, and a brief description of its functionality is provided at the beginning of the file. The overall code follows a typical FEM workflow (element-level procedures, assembly to global matrices, solution for unknown degrees of freedom). 
+
+The analysis generates output data at each load increment, which are stored in the `Saved_Files` folder. It also prints contour plots for damage, local strain, and nonlocal strain at the last load increment, which are stored in the `Images` folder. 
+
+## IV. Model files structure
+Each model is fully defined in two `".txt"` files: a mesh file (`modelName_mesh.txt`) and a parameter file (`modelName_parameters.txt`). The mesh file contains the definitions of:
+
+- Nodal coordinates  
+- Element connectivity
+- Imposed Dirichlet boundary conditions
+
+The parameter file contains several hyperparameters of the problem, and they are presented in detail below (see Section “VIII. Details of parameter file”). <ins>The parameter file must reference the associated mesh file.</ins> This occurs at the last line of the parameter file, which links to the appropriate mesh file.
+
+To facilitate the user’s understanding of the model files, we provide in the `Test Problems` folder the files for 4 sample cases: a Symmetric Single Notch Tension (SSNT) problem with Coarse, Intermediate and Fine mesh resolution (following [5]), and a Double Notch Tension test (termed as Direct Tension Test in [6]). The user should create files that follow the structure of the provided benchmarks.
+
+## V. Running the code
+In order for the user to implement our code on one of the provided benchmark problems (let us assume “SSNT_Coarse”), the following steps are required. The user needs to:
+
+1. Ensure MATLAB and the Parallel Computing Toolbox are installed.
+2. Download the `Code` and `Test Problems` folders.
+3. Open the `SSNT_Coarse_parameter.txt` file and adjust the parameters according to their desired setup (see “VII. Details of parameter files” for details). 
[…]
