+++ Paper/paper.bib
@@ -24,7 +24,7 @@
   journal   = {Applied Mechanics Reviews},
   volume    = {50},
   number    = {10},
-  pages     = {581--591},
+  pages     = {581-591},
   year      = {1997},
   doi       = {10.1115/1.3101670}
 }
@@ -37,28 +37,41 @@
   doi       = {10.1007/s00466-024-02473-5}
 }
 
+@article{mazars1986,
+  author    = {Mazars J.},
+  title     = {A description of micro-and macroscale damage of concrete structures},
+  journal   = {Engineering Fracture Mechanics},
+  volume    = {25},
+  number    = {5-6},
+  pages     = {729-737},
+  year      = {1986},
+  doi       = {10.1016/0013-7944(86)90036-6}
+}
+
+@article{geers1998,
+  author    = {Geers, M.G.D. and de Borst, R. and Brekelmans, W.A.M. and Peerlings, R.H.J.},
+  title     = {Strain-based transient-gradient damage model for failure analyses},
+  journal   = {Computer Methods in Applied Mechanics and Engineering},
+  year      = {1998},
+  doi       = {10.1016/S0045-7825(98)80011-X}
+}
+
 @misc{mathworks2024a,
   title     = {Choose Between Thread-Based and Process-Based Environments},
   year      = {2024},
-  howpublished = {\url{https://www.mathworks.com/help/parallel-computing/choose-between-thread-based-and-process-based-environments.html}}
+  howpublished = {\url{https://www.mathworks.com/help/parallel-computing/
+choose-between-thread-based-and-process-based-environments.html}}
 }
 
-@misc{mathworks2024b,
-  title     = {Compare Performance of parfor, parfeval, and spmd},
-  year      = {2024},
-  howpublished = {\url{https://www.mathworks.com/help/parallel-computing/choose-spmd-parfor-parfeval.html}}
-}
-
-@misc{mathworks2024c,
-  title     = {spmd - Execute code in parallel on workers of parallel pool},
-  year      = {2024},
-  howpublished = {\url{https://www.mathworks.com/help/parallel-computing/spmd.html}}
-}
-
-@misc{mathworks2024d,
-  title     = {parfor – Execute for-loop iterations in parallel on workers},
-  year      = {2024},
-  howpublished = {\url{https://www.mathworks.com/help/parallel-computing/parfor.html}}
+@article{peerlings1998,
+  author    = {Peerlings, R. H. J. and de Borst, R. and Brekelmans, W. A. M. and Geers, M. G. D.},
+  title     = {Gradient-enhanced damage modelling of concrete fracture},
+  journal   = {Mechanics of Cohesive-frictional Materials},
+  volume    = {3},
+  number    = {4},
+  pages     = {323-342},
+  year      = {1998},
+  doi       = {10.1002/(SICI)1099-1484(1998100)3:4<323::AID-CFM51>3.0.CO;2-Z}
 }
 
 @article{pantidis2023,
