+++ Paper/paper.md
@@ -28,42 +28,88 @@ affiliations:
    name: Civil and Urban Engineering Department, New York University Abu Dhabi, UAE
    
 date: 24 October, 2024
+
 bibliography: paper.bib
+nocite: |
+  @*
 ---
 # Statement of Need
-Modeling fracture in materials and structures holds immense importance in our efforts to understand how materials fail and hence design more fracture-resistant structures. Among the several theories developed over the last decades, continuum damage mechanics (CDM) studies the behavior of cracks in materials and structures from the viewpoint of continuous stiffness degradation as the crack propagates inside the domain [@lemaitre2012]. CDM simulations are commonly implemented using the Finite Element Method, however the computational cost is notoriously elevated. Several remedies have been proposed to reduce this computational cost, including domain decomposition approaches [@pebrel2008] and adaptive refinement techniques [@li1997]. Implementing parallel computing methods in CDM codes, where multiple computing resources are simultaneously used, is another promising but significantly less-explored avenue. Compared to the other approaches, very few resources are publicly available on this front, despite parallelization becoming increasingly widespread. With an increasing number of commercial electronic devices with multiple cores and threads, each capable of running its own instruction stream, there remains a major untapped potential for significant reductions in the computational expense. 
-This open source code utilizes parallelization techniques for MATLAB in order to significantly accelerate CDM simulations. Building upon previous work by the authors, we develop a parallel version of the MATLAB code that was introduced in [@saji2024], and we demonstrate the additional efficiency over its serial counterpart. We address the common issue of parallel code portability across different computing devices, and we demonstrate the scaling of computational savings at smaller mesh sizes. The code is intended to be used both by researchers and students working with CDM.
+Modeling fracture in materials and structures holds immense importance in our efforts to understand how materials fail and hence design more fracture-resistant structures. Among the several theories developed over the last decades, continuum damage mechanics (CDM) studies the behavior of cracks in materials and structures from the viewpoint of continuous stiffness degradation as the crack propagates inside the domain [@lemaitre2012]. CDM simulations are commonly implemented using the Finite Element Method, however the associated computational cost is notoriously elevated. Several remedies have been proposed to reduce this cost, including domain decomposition approaches [@pebrel2008] and adaptive refinement techniques [@li1997]. Compared to these approaches, very few resources are publicly available on implementing parallel computing methods in CDM codes, despite parallelization becoming increasingly widespread. 
 
+This open source code utilizes parallelization techniques for MATLAB in order to significantly accelerate CDM simulations. Building upon previous work by the authors [@saji2024], we develop a parallel MATLAB code and demonstrate the additional efficiency over its serial counterpart. The code is geared for quasi-brittle materials, and it is implemented with two relevant damage models (Mazars’ [@mazars1986] and Geers’ model [@geers1998]). Both the Unified arc-length (UAL) and Newton-Raphson solvers can be used. We address the common issue of parallel code portability across different computing devices, and we demonstrate the scaling of computational savings at smaller mesh sizes. 
+
[…]
