+++ Code/FEM_Main_Script_2D_HPC.m
@@ -1,409 +1,0 @@
-tic; % start timer for total code run time
-
-clc
-clear
-close all
-format compact
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%============================ CREATE POOL ================================%
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-% Create a parallel pool with 6 threads
-if isempty(gcp('nocreate'))
-    parpool('Threads',6);
-end
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%============================ USER INPUTS ================================%
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-% Enter location of current file path
-main_file_path='C:\...';                    % Replace with file path
-% Enter location of results storage folder
-save_path='C:\...\Saved_Files';             % Replace with file path 
-
-cd (main_file_path);
-
-% Enter your model name
-model_name = "SSNT_Fine";
-
-% Enter convergence tolerance
-tolerance=1e-8;
-
-% For UAL routine:
-% Enter the upper limit of ArcLength
-ArcLength_upper_limit = 1e-2;
-% Enter the lower limit of ArcLength
-ArcLength_lower_limit = 1e-24;
-
-% I] Local Damage
-% a) Choose the solution scheme to be used:
-Scheme_ID = 1;            % 1 - Partitioned Consistent (PC) scheme   2 - Partitioned Non-Consistent (PNC) scheme
-% b) Choose a strain tolerance (ST) value
-ST = 1e-6;                % Recommended range: 1e-5 to 1e-7
-
-% =========================================================================
-% ============================== IMPORTANT ===============================%
-% =========================================================================
-% ------------------------ Parameters file guide --------------------------
-% -------------------------------------------------------------------------
-% Choose solution scheme
-% SolverID:        1 - Local                (No._DOF_per_node: 2)
-%                  2 - Nonlocal Gradient    (No._DOF_per_node: 3)
-
-% NOTE: Ensure that the "No._DOF_per_node" is updated in the parameter file
-% when switching from Local to Nonlocal Gradient model
-
-% TangentID:       1 - Analytical
-%                  2 - Numerical
-%
-% RoutineID:       1 - Unified Arclength
-%                  2 - Newton Raphson
-%
-% Constraint type: 1 - Cylindrical
-%                  2 - Spherical
-% -------------------------------------------------------------------------
-
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% ======================= INCLUDE GLOBAL VARIABLES ========================
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-func_include_flags;
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
[…]
