+++ Code/FEM_Main_Script_2D.m
@@ -6,51 +6,63 @@ close all
 format compact
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%============================ CREATE POOL ================================%
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-% If no parallel pool created by user, create a parpool with 3 threads.
-if isempty(gcp('nocreate'))
-    parpool('Threads',3); % Adjust number of threads
-end
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %============================ USER INPUTS ================================%
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 % Enter location of current file path
-main_file_path='C:\...';                    % Replace with file path
-% Enter location of results storage folder
-save_path='C:\...\Saved_Files';             % Replace with file path
-
-cd (main_file_path);
+main_file_path='';                   % Enter the file path of the Code folder here
 
 % Enter your model name
-model_name = "SSNT_Coarse";
+model_name = "SSNT_Coarse";          % Enter the mesh name here
+pool_type = 'Threads';               % Type of parallel pool [enter 'Threads' or 'Processes']
+n_threads = 3;                       % No. of parallel threads [Depends on number of cores in the user device] (Default: 3)
 
-% Enter convergence tolerance
-tolerance=1e-8;
-
-% For UAL routine:
-% Enter the upper limit of ArcLength
-ArcLength_upper_limit = 1e-2;
-% Enter the lower limit of ArcLength
-ArcLength_lower_limit = 1e-24;
-
-% I] Local Damage
-% a) Choose the solution scheme to be used:
-Scheme_ID = 1;            % 1 - Partitioned Consistent (PC) scheme   2 - Partitioned Non-Consistent (PNC) scheme
-% b) Choose a strain tolerance (ST) value
-ST = 1e-6;                % Recommended range: 1e-5 to 1e-7
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%=================== NO CHANGES NEEDED AFTER THIS LINE ===================%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+cd (main_file_path);
+% Path of results storage folder
+save_path = strcat(main_file_path,'\Saved Files');            
+% Path of image storage folder
+image_path =strcat(main_file_path,'\Images');            
 
 % =========================================================================
 % ============================== IMPORTANT ===============================%
 % =========================================================================
 % ------------------------ Parameters file guide --------------------------
 % -------------------------------------------------------------------------
+
+% tolerance        : Convergence tolerance
+
 % Choose solution scheme
 % SolverID:        1 - Local                (No._DOF_per_node: 2)
 %                  2 - Nonlocal Gradient    (No._DOF_per_node: 3)
+
+% Choose the solution scheme to be used [UAL local only] :
+% Scheme_ID:       1 - Partitioned Consistent (PC) scheme   
+%                  2 - Partitioned Non-Consistent (PNC) scheme
+
+% max_failed_attempts: Maximum allowed number of failed increments [UAL only]
+
[…]
