### 2025-05-20 a92d4f5
Update func_read_input_file.m
