### 2025-05-20 3eabe5f
Update and rename parallel_func_globalstiffness.m to func_globalstiffness.m
