### 2025-05-20 addb0db
Update func_UnifiedArcLength_DispControl.m
