+++ Code/func_NewtonRaphson_DispControl.m
@@ -1,0 +1,172 @@
+function [J, dofs, Res_F_F_norm, Res_u_norm, iteration, history_var_mat, gausspoints_prop_mat, nodes_prop_mat, Reactions_x, Reactions_y] = func_NewtonRaphson_DispControl(Damage_type,k_damage_parameter,eq_strain_type, RoutineID, dofs, fixnodes_applied, increment, Delastic, history_var_mat, num_elem_at_node, n_hood, weights, tolerance, IsProj, reaction_calc)
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% ==================== NEWTON RAPHSON ANALYSIS METHOD =====================
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+% Include global variables
+func_include_flags;
+
+% -------------------------------------------------------------------------
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% SOLVER ROUTINE %%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% -------------------------------------------------------------------------
+if IsProj == 0
+
+    % Create empty entries for element/nodal properties 
+    gausspoints_prop_mat = [];
+    nodes_prop_mat       = [];   
+
+    % Create copies of the kappa matrix from the end of the previous increment
+    history_var_mat_previousinc = history_var_mat;
+    
+    % Initiate the iteration counter within each loading step
+    iteration = 1;
+    
+    % Evaluate stiffness matrix [K] and history variable matrix at the start of the increment
+    [~, J, DRdu, ~, ~, ~, history_var_mat, ~, ~, ~] = func_globalstiffness(Damage_type,k_damage_parameter,eq_strain_type,dofs,Delastic,history_var_mat_previousinc,num_elem_at_node,n_hood,weights,0,[],IsProj,RoutineID);
+                                                      
+    % Partition K and DRdu
+    if TangentID == 1
+        [~, ~, ~, J_F, ~, ~, listofnodes_ebc, listofnodes_nbc] = func_partitionK(fixnodes_applied, J); 
+    elseif TangentID == 2
+        [~, ~, ~, DRdu_F, ~, ~, listofnodes_ebc, listofnodes_nbc] = func_partitionK(fixnodes_applied, DRdu);
+    else
+        disp("Check your choice of solver!")
+    end
+    
+    % Calculate the fixed displacement increment and partition dofs
+    [~, dofs_E, dofs_F] = func_partitiond_fixnodes(fixnodes_applied,dofs);
+    
+    % Assemble the global dofs
+    dofs(listofnodes_ebc) = dofs_E; 
+    dofs(listofnodes_nbc) = dofs_F;
+    
+    % Calculate and partition the residual force vector (based on internal stresses) after delta_dofs_E has been applied
+    [~, ~, ~, Res_F, ~, ~, ~, ~, ~, ~] = func_globalstiffness(Damage_type,k_damage_parameter,eq_strain_type,dofs,Delastic,history_var_mat_previousinc,num_elem_at_node,n_hood,weights,0,[],IsProj, RoutineID);
+    [~, ~, Res_F_F] = func_partitionf(fixnodes_applied,Res_F); 
+
+    % Compute and store the norm of the residual at the free boundary 
+    Res_F_F_norm_first      = norm(Res_F_F,2);
+    Res_F_F_norm(iteration) = Res_F_F_norm_first;
+    
+    % The while loop runs until either the error is smaller than a specific
[…]
