### 2025-05-20 022c415
Update func_calc_Beta.m
