### 2025-05-20 fd712e7
Update func_elstif_Nonlocgradient.m
