+++ Code/func_elstif_Nonlocintegral.m
@@ -1,247 +1,0 @@
-function [k_el, damage_el, gausspoints_prop_mat_elem, nodes_prop_mat_elem, f_internal_el, f_external_el] = func_elstif_Nonlocintegral(lmn,lmncoord,dofs,Delastic,damage_mat_previousinc,alpha_val,beta_val,e_delta,dmax,n_hood,weights,damage_mat_it,IsM,IsProj)
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% ============ ELEMENT STIFFNESS MATRIX & INTERNAL FORCE VECTOR ===========
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-
-% -------------------------------------------------------------------------
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% SOLVER ROUTINE %%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% -------------------------------------------------------------------------
-if IsProj == 0
-
-    % Create empty entries for element/nodal properties 
-    gausspoints_prop_mat_elem = [];
-    nodes_prop_mat_elem = [];   
-
-    % Unroll dofs into a vector
-    dofs_vec = dofs(:); % 8x1
-    
-    % Initialize matrices and vectors at zero
-    k_el            = zeros(8,8);
-    f_internal_el   = zeros(8,1);
-    f_external_el   = zeros(8,1);
-    damage_el       = zeros(1,4);
-
-    % Setting up: 
-    npoints = func_numberofintegrationpoints;   % Number of integration points
-    xilist  = func_integrationpoints;           % Positions of integration points
-    w       = func_integrationweights;          % Weights of integration points
-    
-    % Loop over the integration points
-    for integ_point = 1:npoints
-    
-        % ================== SHAPE FUNCTIONS AND DERIVATIVES ==================
-        % Compute shape functions && derivatives wrt local coords
-        xi = xilist(:,integ_point);             % Size: 2x1
-        dNdxi = func_shapefunctionderivs(xi);   % Size: 4x2
-        
-        % Compute the jacobian matrix && its determinant
-        dxdxi = lmncoord * dNdxi; % Size: 2x2 - Jacobian matrix
-        dt = det(dxdxi);
-    
-        % Convert shape function derivatives to derivatives wrt global coords
-        dNdx = dNdxi/dxdxi; % Size: 4x2 - Multiplies by inverse of dxdxi
-    
-        % Transpose and expand dNdx (entries of B matrix) to B matrix
-        % dNdx is 4x2 ===> B is 3x8
-        [B, ~] = func_Bmatrix(dNdx);
-    
-    
-        % ==================== LOCAL STRAIN CALCULATIONS ======================
-        % Compute the strain by multiplying the shape function derivatives with
-        % the displacements
-        strain_vec_gxy = B * dofs_vec; % Size: 3x1 (3x8 * 8x1)
-    
-        % Extract the strain values for each integration point
-        exx = strain_vec_gxy(1,1);
-        eyy = strain_vec_gxy(2,1);
-        gxy = strain_vec_gxy(3,1);
-        
-        % Calculate the equivalent strain e_star
-        [e_star,s] = func_estar(exx, eyy, gxy);
-        
-        % Add NaN condition for [s]
-        if isnan(s); s(:) = 0; end
-    
-    
-        % =============== MAZAR DAMAGE MODEL IMPLEMENTATION ===============
[…]
