### 2025-05-20 b61c41e
Delete Code/func_partitionK_FAL.m
