### 2025-05-20 346a76b
Delete Code/func_partitiond_FAL.m
