### 2025-05-20 ab6adc7
Delete Code/func_solve_arclength_eq_FAL.m
