+++ Paper/paper.bib
@@ -38,7 +38,7 @@
 }
 
 @article{mazars1986,
-  author    = {Mazars J.},
+  author    = {Mazars, J.},
   title     = {A description of micro-and macroscale damage of concrete structures},
   journal   = {Engineering Fracture Mechanics},
   volume    = {25},
@@ -57,6 +57,7 @@
 }
 
 @misc{mathworks2024a,
+  author    = {MathWorks},
   title     = {Choose Between Thread-Based and Process-Based Environments},
   year      = {2024},
   howpublished = {\url{https://www.mathworks.com/help/parallel-computing/
