+++ Paper/paper.md
@@ -38,7 +38,7 @@ Modeling fracture in materials and structures holds immense importance in our ef
 
 This open source code utilizes parallelization techniques for MATLAB in order to significantly accelerate CDM simulations. Building upon previous work by the authors [@saji2024], we develop a parallel MATLAB code and demonstrate the additional efficiency over its serial counterpart. The code is geared for quasi-brittle materials, and it is implemented with two relevant damage models (Mazars’ [@mazars1986] and Geers’ model [@geers1998]). Both the Unified arc-length (UAL) and Newton-Raphson solvers can be used. We address the common issue of parallel code portability across different computing devices, and we demonstrate the scaling of computational savings at smaller mesh sizes. 
 
-There are several FEM libraries with parallel capabilities publicly available, such as FEniCS, OOFEM, Akantu, OpenSees, deal.II, but they use different strategies, either domain decomposition or MPI. While such software is extensive and efficient, it requires expertise from the user. Our code is unique in its accessibility given that is written in MATLAB, which many users are familiar with. Also, it features the implementation of a newly developed and robust arc-length solver (UAL) developed in [@saji2024], which has demonstrated its superior performance against the force-controlled arc-length (FAL) and Newton-Raphson (NR) solvers  both in terms of accuracy and time efficiency. 
+There are several FEM libraries with parallel capabilities publicly available, such as FEniCS, OOFEM, Akantu, OpenSees, deal.II, but they use different strategies such as domain decomposition and/or Message Passing Interface (MPI). While such software is extensive and efficient, it requires expertise from the user. Our code is unique in its accessibility given that is written in MATLAB, which many users are familiar with. Also, it features the implementation of a newly developed and robust arc-length solver (UAL) developed in [@saji2024], which has demonstrated its superior performance against the force-controlled arc-length (FAL) and Newton-Raphson (NR) solvers  both in terms of accuracy and time efficiency. 
 
 # Methodology and Software Implementation
 A flowchart of the code’s general functionality is shown in \autoref{fig:general_flowchart}. The function on which we focus is the “func_globalstiffness” function, which calculates and assembles the global stiffness matrix, force vectors, and residual vector. The function also calculates the projection matrices required for contour plotting once the numerical analysis of each load increment is complete.
