### 2025-05-20 Habiba-Eldababy/Parallel-CDM#2 @Habiba-Eldababy
Based on the reviewer’s comment, we have significantly upgraded the documentation of the submission. To this end, we have created a “ParallelCDM_Documentation” file which addresses all the points raised. Overall, this document includes information about the software features and software prerequisites, detailed explanation of the code and files structure, detailed steps on how to run the code, explanation of the hyperparameters, and more.

With regards to each specific point raised in the comment above:

- W.r.t. the first bullet point: Each model is fully defined within a mesh file and a parameter file (both “.txt”), and several sample working examples exist in the folder “Test problems”. The steps to run the code are presented in the Documentation file.
    - Regarding the suggestion for factorization: we have increased the code efficiency by cleaning up several files, and moving all user-defined variables in the parameter “.txt” file. To run the code, the user now only needs to edit the file path, model name, and parallel pool in the main script. This is indicated by relevant comments in the documentation.

- W.r.t. the second bullet point: The mesh-related files include definitions of nodal coordinates, element connectivity, and imposed Dirichlet boundary conditions. The parameter file includes the parameters of the problem. Though not restricting, we suggest the creation of the mesh files with the GMSH software. Our code supports both structured and unstructured meshes. The elements should be quadrilateral elements with 4 Gauss (integration) points.

- W.r.t. the third bullet point: The model hyperparameters are fully documented in section “VII. Details of parameter files”. A detailed investigation of the impact of each parameter is considered out-of-scope for this submission, as it would significantly exceed the acceptable length of the article. Instead, in this submission we mainly explore the impact of parameters which are pertinent to the parallelization (threads, processes, etc.)

- W.r.t. the fourth bullet point: The code outputs a data file stored in the “Saved_Files” folder at each increment as well as contour plots for damage, local strain, and nonlocal strain at the last load increment stored in the “Images” folder.

This information is explained in the documentation so users are able to interact with the code and potentially modify it to their needs.
