### 2025-05-20 Habiba-Eldababy/Parallel-CDM#3 @Habiba-Eldababy
We thank the reviewer for this comment which increases the reliability of our code. We have tested our code with the direct tension test problem with L=50mm configuration reported by Peerlings et al. and reported the result in the paper in the section “Testing and Results”. The code and the solution in Peerlings et al. show the results are in excellent agreement, an observation which increases the robustness and trustworthiness of our code.
