### 2025-05-20 #7610 @Habiba-Eldababy
@vijaysm @tiburoch @diehlpk
We would like to warmly thank both reviewers for their careful revision and valuable feedback, which helped to improve the quality of our submission. Based on their comments we have significantly re-structured and enhanced our contribution. A detailed [point-by-point response](https://github.com/Habiba-Eldababy/Parallel-CDM/blob/79c7a137a3544c1242565c81e7fa8bec9565c888/JOSS_Response%20to%20reviewers.pdf) and explanation of the modifications made is provided.
