### 2025-05-20 #7610 @Habiba-Eldababy
> I second many of the review points that [@tiburoch](https://github.com/tiburoch) has already raised. The repository is missing several recommended and required practices.
>
> 1. community contribution guidelines,
> 2. documentation related to additional dependencies (parallel computing toolbox) and pointers to tools for generating new meshes for test cases,
> 3. detailed documentation describing the test suite and explaining functionality using tutorials/examples
> 4. description of drivers FEM_Main_Script_2D.m and FEM_Main_Script_2D_HPC.m, their usage, and along with the differences between the two versions
> 5. documentation detailing the control of process vs thread-based parallelized runs and methods to change the number of tasks (pointers to MATLAB documentation may suffice along with instructions to run the driver)
>
> I also have a general comment about the paper.
>
> * In Figure 1., what happens when you update the parameters? There is perhaps an arrow missing to restart the iterations?
> * In Figure 2., the solver workflow, I only see the assembly of the stiffness matrix being done in parallel. Does the solver use any parallelization? This is not indicated in the flowchart.
> * For performance results in Figures 5 and 6, it may be useful to consider your "medium" mesh as the coarse one, and use another level of refinement for the fine one to showcase better results on the workstation and the cluster.
> * Is the parallel efficiency of MATLAB thread-level parallelism limited? The scaling plots plateau immediately beyond 2-3 threads on the workstation and 8 threads on a cluster. While there is a reduction in total runtime compared to serial execution, does the author think that leveraging MATLAB's thread-level parallelism provides a speedup comparable to domain-decomposition-based approaches? You can discuss these implications as part of your conclusions.
>
> Please address these suggestions and let me know when the repository/paper are ready to be reviewed again. [@Habiba-Eldababy](https://github.com/Habiba-Eldababy) [@diehlpk](https://github.com/diehlpk)

We would like to thank the reviewer for the attentive review and thoughtful questions regarding the parallelization. We have addressed all comments raised, and below we provide our response:

•	**W.r.t. point 1:** the repository has now been updated to include the required and recommended practices mentioned. Community contribution is most welcome and we have indicated how this can be done in the README and Documentation file.
•	**W.r.t. point 2:** Information on the software prerequisites (MATLAB and the Parallel Computing Toolbox) and mesh-generation software (such as GMSH) is included in the pertinent sections of the Documentation file.
•	**W.r.t. point 3:** A test for our code is available and detailed in the paper, along with working examples ready for use in the folder “Test problems”. Additionally, with regards to the documentation aspect of the question, we note that each code function includes an opening comment that describes its role, in order to facilitate user interaction and future modifications.
•	**W.r.t. point 4:** Following the re-structuring of our code, we note that now there is only one “FEM_Main_Script_2D.m” file. The user can use this file alone to run the code on any device (local or HPC), with parallel computing capability, by simply modifying the number of cores/threads to be used in the main script.
•	**W.r.t. point 5:** In the revised version of the code we describe the parallel pool and number options in section “VIII. Tips for testing and improved performance” of the Documentation file. In addition, the user will encounter the option to modify the pool in the user input area of the “FEM_Main_Script_2D.m” file.

We also appreciate the reviewer’s thoughtful comments on the paper and address them below one-by-one.
•	There was indeed a missing - this has been updated to indicate that the iterations restart. The flowcharts have also been edited to increase readability.
•	The solver itself does not use any parallelization – hence we do not make a statement about that in the workflow.
•	On the front of using a finer mesh, we underline that we are currently limited by memory constraints that arise when running models with more elements than those reported. We believe our mesh discretization scheme, which leads up to 10,201 elements, is sufficient to demonstrate the trend for progressive speedup (at least 3 times faster).
•	The plateauing behavior observed beyond 2–3 threads on the workstation and 8–16 threads on the HPC cluster is primarily due to the overhead costs inherent to thread-based parallelism in MATLAB. These are related to the worker setup, memory synchronization, and data sharing overhead, which begin to outweigh the benefits as thread count increases without a proportionally larger problem size. These findings are illustrated in Fig. 7 and Fig. S2 of our paper. With regards to its comparison to domain decomposition methods, we acknowledge that such methods tend to show stronger trends of scalability for large-scale problems solved on distributed memory systems. Libraries like PETSc, deal.II, and OpenSees are designed to work with MPI-based parallelism and are well-suited to large-scale simulations. However, domain decomposition comes with considerable implementation complexity and typically requires users to have expertise in low-level parallel programming and mesh partitioning. In contrast, our code adopts MATLAB's thread-based shared-memory model, which offers a practical and user-friendly approach to accelerate small- to medium-sized CDM problems. Despite the inherent limitations of thread-level parallelism, we demonstrate that our implementation provides 3–10× speedup over serial execution, particularly when the mesh size is sufficient and memory is not a limiting factor.
