### 2025-06-17 c65bbd1
Update paper.md
Added software references, solver comparison, and modified figure captions to indicate UAL solver usage.
