### 2025-06-17 de0e27f
Update paper.bib
