+++ Paper/paper.bib
@@ -38,7 +38,7 @@
 }
 
 @article{mazars1986,
-  author    = {Mazars, J.},
+  author    = {Mazars J.},
   title     = {A description of micro-and macroscale damage of concrete structures},
   journal   = {Engineering Fracture Mechanics},
   volume    = {25},
@@ -56,8 +56,83 @@
   doi       = {10.1016/S0045-7825(98)80011-X}
 }
 
+@software{fenics2024,
+  title = {FEniCS Project},
+  author = {{The FEniCS Project}},
+  year = {2024},
+  url = {https://fenicsproject.org/}
+}
+
+@software{mikael_ohman_2020_4339630,
+  author       = {Mikael Öhman and
+                  Borek Patzak and
+                  Jim Brouzoulis and
+                  Vit Smilauer and
+                  milanjirasek and
+                  Peter Grassl and
+                  graspel and
+                  feymark and
+                  CarlSandstrom and
+                  nitramkaroh and
+                  MartinFagerstrom and
+                  pedroskop and
+                  Johannes Främby and
+                  eudoxos and
+                  Adam Sciegaj and
+                  editaDvorakova and
+                  Stanislav Sulc and
+                  Jan Stránský and
+                  karelmikes and
+                  Kristoffer Carlsson and
+                  sancho-s and
+                  vmonkey},
+  title        = {oofem/oofem: OOFEM, version 2.5},
+  month        = dec,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v2.5},
+  doi          = {10.5281/zenodo.4339630},
+  url          = {https://doi.org/10.5281/zenodo.4339630},
+}
+
+@manual{akantu2024,
+  title = {Akantu: A Finite Element Library for Complex Multiphysics Simulations},
+  author = {{EPFL LSMS}},
+  year = {2024},
+  url = {https://akantu.ch/}
+}
+
+@article{Richart2024, doi = {10.21105/joss.05253}, 
+url = {https://doi.org/10.21105/joss.05253}, 
+year = {2024}, 
+publisher = {The Open Journal}, 
+volume = {9}, 
+number = {94}, 
+pages = {5253}, 
+author = {Nicolas Richart and Guillaume Anciaux and Emil Gallyamov and Lucas Frérot and David Kammer and Mohit Pundir and Marco Vocialta and Aurelia Cuba Ramos and Mauro Corrado and Philip Müller and Fabian Barras and Shenghan Zhang and Roxane Ferry and Shad Durussel and Jean-François Molinari},
+title = {Akantu: an HPC finite-element library for contact and dynamic fracture simulations},
+journal = {Journal of Open Source Software} 
+}
+
+@manual{opensees2025,
+  title = {OpenSees: Open System for Earthquake Engineering Simulation},
+  author = {{Pacific Earthquake Engineering Research Center}},
+  year = {2025},
+  url = {https://opensees.berkeley.edu/}
+}
+
+@article{Arndt_The_deal_II_finite_2021,
+author = {Arndt, Daniel and Bangerth, Wolfgang and Davydov, Denis and Heister, Timo and Heltai, Luca and Kronbichler, Martin and Maier, Matthias and Pelteret, Jean-Paul and Turcksin, Bruno and Wells, David},
+doi = {10.1016/j.camwa.2020.02.022},
+journal = {Computers \& Mathematics with Applications},
+month = jan,
+pages = {407--422},
+title = {{The deal.II finite element library: Design, features, and insights}},
+volume = {81},
+year = {2021}
+}
+
 @misc{mathworks2024a,
-  author    = {MathWorks},
   title     = {Choose Between Thread-Based and Process-Based Environments},
   year      = {2024},
   howpublished = {\url{https://www.mathworks.com/help/parallel-computing/
