+++ Paper/paper.md
@@ -38,7 +38,7 @@ Modeling fracture in materials and structures holds immense importance in our ef
 
 This open source code utilizes parallelization techniques for MATLAB in order to significantly accelerate CDM simulations. Building upon previous work by the authors [@saji2024], we develop a parallel MATLAB code and demonstrate the additional efficiency over its serial counterpart. The code is geared for quasi-brittle materials, and it is implemented with two relevant damage models (Mazars’ [@mazars1986] and Geers’ model [@geers1998]). Both the Unified arc-length (UAL) and Newton-Raphson solvers can be used. We address the common issue of parallel code portability across different computing devices, and we demonstrate the scaling of computational savings at smaller mesh sizes. 
 
-There are several FEM libraries with parallel capabilities publicly available, such as FEniCS [@fenics2024], OOFEM [@mikael_ohman_2020_4339630], Akantu [@Richart2024], OpenSees [@opensees2025], and deal.II [@Arndt_The_deal_II_finite_2021], but they use different strategies such as domain decomposition and/or Message Passing Interface (MPI). Our code is unique in its accessibility given that is written in MATLAB, which many users are familiar with. Also, it features the implementation of a newly developed and robust arc-length solver (UAL) developed in [@saji2024], which has demonstrated its superior performance against the force-controlled arc-length (FAL) and Newton-Raphson (NR) solvers both in terms of accuracy and time efficiency. UAL can trace equilibrium paths with snap-backs, making it a suitable solver for a wide range of scenarios, which NR cannot capture. Furthermore, UAL is not constrained by the need for small step sizes and relaxed convergence tolerances that limit the usage of the FAL solver. The behavior of the three solvers is compared and shown in \autoref{fig:Solvers_comparison}, for a 1D bar problem modeled using the non-local gradient damage and a weakened region in the middle (modulus of elasticity is reduced by a factor of 2)[@saji2024]. The UAL solver models the entire non-linear equilibrium path in 0.98 seconds with 97 increments, while the FAL and NR solvers require 1–2 orders of magnitude more time and increments to trace the same path with a more relaxed convergence tolerance [@saji2024].
+There are several FEM libraries with parallel capabilities publicly available, such as FEniCS [@fenics2024], OOFEM [@mikael_ohman_2020_4339630], Akantu [@Richart2024], OpenSees [@opensees2025], and deal.II [@Arndt_The_deal_II_finite_2021], but they use different strategies such as domain decomposition and/or Message Passing Interface (MPI). Our code is unique in its accessibility given that is written in MATLAB, which many users are familiar with. Also, it features the implementation of a newly developed and robust arc-length solver (UAL) developed in [@saji2024], which has demonstrated its superior performance against the force-controlled arc-length (FAL) and Newton-Raphson (NR) solvers both in terms of accuracy and time efficiency. UAL can trace equilibrium paths with snap-backs, making it a suitable solver for a wide range of scenarios, which NR cannot capture. Furthermore, UAL is not constrained by the need for small step sizes and relaxed convergence tolerances that limit the usage of the FAL solver. The behavior of the three solvers is compared and shown in \autoref{fig:Solvers_comparison}, for a 1D bar problem modeled using the non-local gradient damage and a weakened region in the middle (modulus of elasticity is reduced by a factor of 2)[@saji2024]. The UAL solver models the entire non-linear equilibrium path in 0.98 seconds with 97 increments, while the FAL and NR solvers require 1–2 orders of magnitude more time and increments to trace the same path with a more relaxed convergence tolerance [@saji2024]. For all results presented in the following sections we use the UAL solver.
 
 ![Force-displacement plot and time, increment, and tolerance comparison using the UAL, FAL, and NR solvers\label{fig:Solvers_comparison}](force_displacement_solver_comparison.png){width=75%}
 
