### 2025-06-18 Habiba-Eldababy/Parallel-CDM#3 @tiburoch
@Habiba-Eldababy and thank you for adding a test in the manuscript and the corresponding input file in the code source.

However, there is a minor misunderstanding regarding the test requirements. The objective is that anyone acquiring the software or developing in it should be able to test the capabilities of the software and verify that the core functionalities are working. Say for example that I decide to extend your code to 3D, I need to make sure I did not alter the previous functionality of the software and this is done through tests that compare the output of a simulation to a reference solution (either analytical or from a version of the software you know is correct).

The direct tension problem that you included as an input file is a good start, but what is missing is the comparison with a reference solution directly inside the software. I suggest that the direct traction test script should be separated from the examples, and should include the computation of the error between the script output and a reference solution (for example you can compute the error of the force displacement curve shown in Fig.4). Based on the value of the error, the script should return a message stating if the test is passed or not.

Ideally, your code should also include some unit tests (i.e., tests that only evaluate the behavior of a single function) for the basic functionalities of your code (for example, assembling the stiffness matrix, the various damage laws, etc... ).

Cheers,
Thibault
