### 2025-06-25 #7610 @Habiba-Eldababy
> Hi [@Habiba-Eldababy](https://github.com/Habiba-Eldababy), thank you for tackling my last comments. The repository has been greatly improved since the submission, thank you for your and your colleagues' efforts.
>
> There are still a few minor points to be improved in the repository. I have added a few small comments regarding the tests and the examples directly in the repository ([Habiba-Eldababy/Parallel-CDM#3](https://github.com/Habiba-Eldababy/Parallel-CDM/issues/3), [Habiba-Eldababy/Parallel-CDM#2](https://github.com/Habiba-Eldababy/Parallel-CDM/issues/2)).
>
> Regarding the paper itself, I have a question directly for you [@diehlpk](https://github.com/diehlpk) . The paper is currently quite long, with many results on the execution time. This includes comparisons between serial/parallel thread-based/parallel process-based, various mesh sizes, and platform type (laptop, workstation, HPC). My understanding of the JOSS guidelines is that the paper itself should be relatively short and only include necessary information on the software performance, but the guidelines are not very specific on how much detail can/should be included. Do you have any comments on this [@diehlpk](https://github.com/diehlpk) ?
>
> In my opinion, the paper could be significantly shortened while still retaining the relevant points regarding the code and its performance. Maybe [@vijaysm](https://github.com/vijaysm), you can also comment on this before we ask [@Habiba-Eldababy](https://github.com/Habiba-Eldababy) to make significant changes to the paper itself.
>
> If we agree that the paper can be shortened, my suggestion would be to only include the current Fig8 showing the scaling capabilities of the code on HPC with the finest mesh size (and remove Fig6 and Fig7). The points raised with Fig6 and Fig7, such as process-based parallelism not being efficient in the case of this software, could be included as a comment in the text.
>
> I will give more specific comments on the paper once we agreed on shortening it (or not).
>
> Best regards, Thibault

Hi @tiburoch, Thank you for your reply and attentive comments. I will work on the two new points you made regarding the tests and minor additions to the documentation. As for shortening the paper, I see that it is quite long compared to other JOSS papers.

If I remove Fig. 6 and Fig. 7 then there would be no laptop/workstation results. Is this what you suggest? Also, would you like me to keep the newer figures added during the review process (Fig. 1 and Fig. 4)? I can also move the supplementary information section at the end of the paper to the documentation.

Overall, if you could provide me with more details as you mentioned that would be very helpful to avoid many rounds of revisions.

Best regards,
Habiba
