### 2025-06-26 #7610 @diehlpk
@Habiba-Eldababy I would recommend to remove **Supplementary Information** too.
