### 2025-07-08 #7610 @Habiba-Eldababy
@tiburoch @diehlpk @vijaysm

Thank you for your latest suggestions and comments. We have significantly shortened the paper to fit the JOSS guidelines. We now show only the HPC result in the paper, and other results are in the documentation. The conclusion was removed too, and the supplementary results were also moved to the documentation.

In terms of the test code https://github.com/Habiba-Eldababy/Parallel-CDM/issues/3 , we added "Direct_Tension_Software_Test.m" which compares a current Direct Tension Test run of the code to reference F-D data, calculates the error (max force error and L2 norm), then tells the user if the test is passed or failed.

https://github.com/Habiba-Eldababy/Parallel-CDM/issues/2 The documentation now includes more details about the two examples and the mesh sizes.

https://github.com/Habiba-Eldababy/Parallel-CDM/issues/1 I believe the comments in this issue have been addressed previously in the paper, and to the end of shortening the paper, points 2 and 3 are now addressed in the documentation instead.

Thank you,
Habiba
