### 2025-07-17 #7610 @vijaysm
@Habiba-Eldababy Please update the repository and address the following comments.

- Provide a link to [ParallelCDM_Documentation.md](https://github.com/Habiba-Eldababy/Parallel-CDM/blob/main/Documentation/ParallelCDM_Documentation.md) from the main README file.
- I suggest splitting [ParallelCDM_Documentation.md](https://github.com/Habiba-Eldababy/Parallel-CDM/blob/main/Documentation/ParallelCDM_Documentation.md) and moving the following sections to README directly to provide new users an immediate bird's eye view of both features and requirements for using the software
    1. Software Features
    2. Software Prerequisites
    3. Code structure
- Describe how users may contribute to the code and collaboratively develop the software. This is provided in [ParallelCDM_Documentation.md](https://github.com/Habiba-Eldababy/Parallel-CDM/blob/main/Documentation/ParallelCDM_Documentation.md) under "Contributions, Support, and Reporting Issues" section, but it might be better to move this into a separate file (Contributing.md) to be placed under the repository root
- The newer version of the manuscript reads much better than before.
