### 2025-07-17 #7610 @tiburoch
Dear @Habiba-Eldababy,

Thank you for the last modifications. I have a few comments on the paper before I can give my approval for publication:

- The statement of need still needs a bit of work. You write that there are several remedies for reducing the cost of CDM simulations (domain decomposition and adaptive refinement techniques), and then say that, compared to these approaches, there are very few parallel codes for CDM.
      1) Does this mean that many publicly available codes implement domain decomposition and adaptive refinement techniques for CDM ? If so, please provide some reference to public packages. If not, maybe it needs to be reformulated.
      2) A few lines after, you list several FEM libraries with parallel capabilities. It may seem contradictory to say that there are very few parallel codes for CDM in this case, as you did a few sentences before. Maybe you want to emphasize the fact that there are no other parallel Matlab codes(?) , or that you have a more efficient way to solve the problem instead ? No need to go into details in the first paragraph, as it should be aimed at a broad non-specialist audience. You could say something along the line of: parallel computing is another way to tackle the cost of CDM computation, and that your code is parallelized AND implements a unique, efficient solver AND is accessible to non-coding specialists as it is written in MATLAB.

In the third paragraph, when you talk about the UAL solver, I think you should still mention how much UAL speeds up compared to Newton-Raphson (in a table or directly in the text, as previously mentioned). You could add it at the end of this sentence: "which has demonstrated its superior performance against the force-controlled arc-length (FAL) and Newton-Raphson (NR) solvers both in terms of accuracy and time efficiency."

Best,
Thibault
