### 2025-07-17 #7610 @Habiba-Eldababy
@tiburoch Thank you for your comments. We’ve revised the “Statement of Need” to avoid confusion and to emphasize that our code is unique in combining parallelization, a robust solver, and accessibility via MATLAB. Additionally, we’ve included the quantitative comparison of UAL to FAL and NR solvers in the text, as you suggested, to demonstrate the efficiency gains.

We hope these changes address your concerns.
