### 2025-08-05 #7610 @vijaysm
Was able to successfully run on NERSC and tested the SSNT_Coarse case with 3 and 6 threads. Produced the following image. I am fine with the current status.
