### 2025-08-10 #7610 @Habiba-Eldababy
Hi @diehlpk, I have completed the above.

1. The author list is verified.
2. The release version number is **v1.0.0** [link](https://github.com/Habiba-Eldababy/Parallel-CDM/releases/tag/v1.0.0)
3. The Zenodo DOI: **10.5281/zenodo.16790622**
4. The title and author list match
5. The license in both is the MIT license

Let me know if there is anything else I should do
