### 2025-08-11 #7610 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
