### 2025-08-13 #7610 @Habiba-Eldababy
Hi @kyleniemeyer, I have updated the paper and archive titles and reviewed & merged the pull request. Let me know if you have any more issues I can resolve.
